fullcircle: Try building from pfunct --compile and check types
With this and running like:

  $ find ~/git/build/v5.1-rc4+/fs/ -name "*.o" | grep -v .mod.o | \
	while read obj ; do
		echo $obj ;
		fullcircle $obj ;
	done

The vast majority of the kernel single compilation unit objects get
the source code for its function prototypes and types used rebuilt,
recompiled and then the original DWARF can be compared with the one
generated from the regenerated C code.

More work needed, but should be a good start and has already helped to
fix several issues, a reported in the previous csets.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>