dwarves: Move BTF loader ahead of the CTF one
When we call the tools without specifying a file format, all the
available format loaders are called, in sequence, till we find the
relevant information, now that we support raw BTF we better move it in
front of the CTF one, as it is way more common in the Linux community.

With this we will not see this warning anymore:

  $ pahole -C list_head  /sys/kernel/btf/vmlinux
  ctf__new: cannot get elf header.
  struct list_head {
  	struct list_head *         next;                 /*     0     8 */
  	struct list_head *         prev;                 /*     8     8 */

  	/* size: 16, cachelines: 1, members: 2 */
  	/* last cacheline: 16 bytes */
  };
  $

That warning has to go, so that other formats, if present after the CTF
one, can be tried.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>