pahole: Check if the type with arguments is present in the current CU
In multi CU (Compile Unit) objects, namely DWARF ones, we go on
processing all of them looking for the types of interest (the ones in
-C), but before we check if the type is in that CU, we first parse its
arguments, when present.

Do that lookup sooner, right after we get hold of the type name, to
avoid needlessly parsing its arguments, now when we run;

  $ pahole -V ~/bin/perf --seek_bytes=0x130 -C 'perf_event_header(sizeof=size,type=type,type_enum=perf_event_type)' --count 26
  pahole: sizeof_operator for 'perf_event_header' is 'size'
  pahole: type member for 'perf_event_header' is 'type'
  pahole: type enum for 'perf_event_header' is 'perf_event_type'
  struct perf_event_header {
  	__u32                      type;                 /*     0     4 */
  	__u16                      misc;                 /*     4     2 */
  	__u16                      size;                 /*     6     2 */

  	/* size: 8, cachelines: 1, members: 3 */
  	/* last cacheline: 8 bytes */
  };
  $

We'll see those verbose messages just once, in the first CU that has the
definitions for the types of interest.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>