pahole: Add support for referencing header variables when pretty printing
To know from where to start reading some variable sized type we need
some information from a domain specific tool such as 'perf report -D'
for perf.data files, i.e.:

  $ perf report -D -i perf.data
  # To display the perf.data header info, please use --header/--header-only options.
  #

  0x130 [0x20]: event: 79
  .
  . ... raw event: size 32 bytes
  .  0000:  4f 00 00 00 00 00 20 00 1f 00 00 00 00 00 00 00  O..... .........
  .  0010:  31 30 9b 3c 00 00 00 00 2e 53 f8 0c 52 8c 01 00  10.<.....S<F8>.R...

  0 0x130 [0x20]: PERF_RECORD_TIME_CONV: unhandled!

So we see that 0x130 is where the first PERF_RECORD_* event is located,
and we can use that with --seek_bytes and with some extra info we can
decode variable sized records:

  $ pahole --seek_bytes=0x130 -C 'perf_event_header(sizeof=size,type=type,type_enum=perf_event_type)' --skip 2 --count 3 ~/bin/perf < perf.data
  {
  	.type = 0x4a,
  	.misc = 0,
  	.size = 0x20,
  },
  {
  	.header = {
  		.type = PERF_RECORD_CGROUP,
  		.misc = 0,
  		.size = 0x28,
  	},
  	.id = 0x1,
  	.path = "/",
  },
  {
  	.header = {
  		.type = PERF_RECORD_CGROUP,
  		.misc = 0,
  		.size = 0x30,
  	},
  	.id = 0x611,
  	.path = "/system.slice",
  },
  $

But if we decode the perf.data file header:

  $ pahole ~/bin/perf -C perf_file_header --count 1 < perf.data
  {
  	.magic = 0x32454c4946524550,
  	.size = 0x68,
  	.attr_size = 0x88,
  	.attrs = {
  		.offset = 0xa8,
  		.size = 0x88,
  	},
  	.data = {
  		.offset = 0x130,
  		.size = 0x588,
  	},
  	.event_types = {
  		.offset = 0,
  		.size = 0,
  	},
  	.adds_features = { 0x16717ffc, 0, 0, 0 },
  },
  $

We see that that 0x130 offset is at field perf_file_header->data.offset,
so lets automate this so that we can try to extract that value and then
use it with --seek_bytes:

  $ pahole --header=perf_file_header --seek_bytes='$header.data.offset' -C 'perf_event_header(sizeof=size,type=type,type_enum=perf_event_type)' --skip 2 --count 3 ~/bin/perf < perf.data
  pahole: the type enum 'perf_event_type' wasn't found in 'util/header.c'
  $

The problem here is that pahole tries to avoid processing all the CUs
(compile units) in a binary, as, so far, it only needed to process one
main type at a time, i.e. the ones in -C/--class.

Now we need multiple types, in the above example we need:

  struct perf_event_header
  struct pref_file_header
  enum perf_event_type

And in this case, the perf binary doesn't have any object/CU that has
all these three types.

To see if the code works we can resort to using BTF, that combines all
types into just one "CU", deduplicating them in the process, so:

  $ pahole --btf_encode ~/bin/perf

And now it works:

  $ pahole -V -F btf --header=perf_file_header --seek_bytes='$header.data.offset' -C 'perf_event_header(sizeof=size,type=type,type_enum=perf_event_type)' --skip 2 --count 3 ~/bin/perf < perf.data
  pahole: sizeof_operator for 'perf_event_header' is 'size'
  pahole: type member for 'perf_event_header' is 'type'
  pahole: type enum for 'perf_event_header' is 'perf_event_type'
  pahole: seek bytes evaluated from --seek_bytes=$header.data.offset is 0x130
  {
  	.type = 0x4a,
  	.misc = 0,
  	.size = 0x20,
  },
  {
  	.header = {
  		.type = PERF_RECORD_CGROUP,
  		.misc = 0,
  		.size = 0x28,
  	},
  	.id = 0x1,
  	.path = "/",
  },
  {
  	.header = {
  		.type = PERF_RECORD_CGROUP,
  		.misc = 0,
  		.size = 0x30,
  	},
  	.id = 0x611,
  	.path = "/system.slice",
  },
  $

In the next csets a fallback approach will allow for this to work even
with DWARF, when we'll notice that stdin wasn't consumed and thus we
need to search for the types needed with cus__find_struct_by_name() &
friends.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>