pahole: Fixup the --class_name parsing wrt class args (type=, sizeof=, etc)
We were going past the end of the argument when the last one had (),
i.e.

   -C 'perf_event_header(sizeof,type,type_enum=perf_event_type)'

We were going after the ending '\0' because when looking for the ',' we
first found the inner ones, as we looked at the first ) from the
current class, we needed to use as the separator the ')' location + one,
but when we have just one a() style class we would put the '\0' not on
top of a ',' but on the already existing one, then increment the
separator to after the '\0', oops, fix it.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>