pahole: Fallback to pretty printing using types in multiple CUs
For DWARF we may end up with a set of arguments that reference multiple
types and those may not be all present in any of the CUs when using
DWARF, i.e. for instance this wasn't possible with DWARF:

  $ pahole ~/bin/perf --header=perf_file_header \
		      --seek_bytes='$header.data.offset' -C
		      'perf_event_header(sizeof,type,type_enum=perf_event_type)'
		      --size_bytes='$header.data.size' < perf.data

As these types are present in a single CU in the DWARF sections of ~/bin/perf:

  'struct perf_file_header'
  'struct perf_event_header'
  'enum perf_event_type'

This works with BTF as we have everything in just one place, one 'CU':

Running the above command line with -F btf and with -F dwarf produces
the same output.

  $ pahole -F dwarf ~/bin/perf --header=perf_file_header --seek_bytes='$header.data.offset' -C 'perf_event_header(sizeof,type,type_enum=perf_event_type)' --size_bytes='$header.data.size'  < perf.data  > dwarf
  $
  $ pahole -F btf ~/bin/perf --header=perf_file_header --seek_bytes='$header.data.offset' -C 'perf_event_header(sizeof,type,type_enum=perf_event_type)' --size_bytes='$header.data.size'  < perf.data  > btf
  $
  $ diff -u btf dwarf
  $
  $ tail btf
  {
  	.type = PERF_RECORD_EXIT,
  	.misc = 0,
  	.size = 48,
  },
  {
  	.type = 68,
  	.misc = 0,
  	.size = 8,
  },
  $ wc btf
   246  642 3863 btf
  $

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>