pahole: Optimize --header processing by keeping the first successfull instance
Since we store both the tag and the cu when we find it, defer deleting
it till we're completely sure we don't need it anymore.

Sometimes this is the only reason for us to fallback to looking at all
the cus to get all the needed types, so keeping it may make the DWARF
"stealer" to be able to, after finding the header type in one CU, find
the other needed types in another and we end up not having to process
all the CUs in a DWARF based session, speeding up the process.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>