pahole: Cache the type_enum lookups into struct enumerator
I.e. when we get the type= field value, look it up in the type_enum=,
get the struct enumerator and if it is the first time that we're looking
up the string representation of that enumerator, do it and then cache
the results, so that next time we reuse it.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>