dwarf_loader: Add minimal handling of DW_TAG_subrange_type
This was found in a ADA object, part of gdb's test suite, for now just
make the code a bit more robust when not finding a type for some struct
member, etc, which avoids segfaults and produces output from ADA
objects, but there are other problems to solve as this is a _type tag, I
need to provide some better support so that type resolution works.

  [foo.debug.gz](https://github.com/acmel/dwarves/files/5257332/foo.debug.gz)

  Foo.debug, an ada exec:
  ```
  $ ~/dwarves/build/pahole foo.debug
  die__process_unit: DW_TAG_subrange_type (0x21) @ <0x10b> not handled!
  die__process_unit: DW_TAG_subrange_type (0x21) @ <0x134> not handled!
  die__process_unit: DW_TAG_subrange_type (0x21) @ <0x148> not handled!
  die__process_class: DW_TAG_subrange_type (0x21) @ <0x201> not handled!
  Segmentation fault (core dumped)
  $

These are fixed, the warnings continue to be produced.

Reported-by: Tom de Vries
Bugtracker: https://github.com/acmel/dwarves/issues/9#issuecomment-696282005
Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>